requirements.txt
pandas
scikit-learn
requests
joblib
python-dotenv
spacy
tiktoken